blog_api_client/backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20230430
dj-rest-auth==4.0.1
Django==4.2.3
django-ckeditor==6.6.1
django-cors-headers==4.1.0
django-debug-toolbar==4.1.0
django-grappelli==3.0.6
django-import-export==3.2.0
django-js-asset==2.1.0
djangorestframework==3.14.0
drf-yasg==1.21.6
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.3
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.1
Pillow==10.0.0
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
xlrd==2.0.1
xlwt==1.3.0
